+++ uploadcare_client/lib/src/transformations/image.dart
@@ -17,9 +17,11 @@ enum ImageFormatTValue {
 }
 
 /// Convert an image to one of the supported output formats: [ImageFormatTValue]
+///
+/// See https://uploadcare.com/docs/transformations/image/compression/#operation-format
 class ImageFormatTransformation extends EnumTransformation<ImageFormatTValue>
     implements ImageTransformation {
-  ImageFormatTransformation(ImageFormatTValue value) : super(value);
+  const ImageFormatTransformation(ImageFormatTValue value) : super(value);
 
   @override
   String get valueAsString {
@@ -43,18 +45,22 @@ class ImageFormatTransformation extends EnumTransformation<ImageFormatTValue>
 
 /// Convert JPEG to Progressive JPEG. Has no effect on non-JPEGs. Does not force image formats to jpeg.
 /// if [value] is equal `true` then are using multi-scan rendering.
+///
+/// See https://uploadcare.com/docs/transformations/image/compression/#operation-progressive
 class ProgressiveTransformation extends BooleanTransformation
     implements ImageTransformation {
-  ProgressiveTransformation([bool value = false]) : super(value);
+  const ProgressiveTransformation([bool value = false]) : super(value);
 
   @override
   String get operation => 'progressive';
 }
 
 /// Instruct Uploadcare CDN whether it should rotate an image according to the EXIF orientation tag or not.
+///
+/// See https://uploadcare.com/docs/transformations/image/rotate-flip/#operation-autorotate
 class AutoRotateTransformation extends BooleanTransformation
     implements ImageTransformation {
-  AutoRotateTransformation([bool value = true]) : super(value);
+  const AutoRotateTransformation([bool value = true]) : super(value);
 
   @override
   String get operation => 'autorotate';
@@ -62,9 +68,11 @@ class AutoRotateTransformation extends BooleanTransformation
 
 /// Rotate an image counterclockwise.
 /// [value] should be in `-360..360` range
+///
+/// See https://uploadcare.com/docs/transformations/image/rotate-flip/#operation-rotate
 class RotateTransformation extends IntTransformation
     implements ImageTransformation {
-  RotateTransformation(int value)
+  const RotateTransformation(int value)
       : assert(value >= -360 && value <= 360, 'Should be in -360..360 range'),
         super(value);
 
@@ -73,6 +81,8 @@ class RotateTransformation extends IntTransformation
 }
 
 /// Flip an image (mirror-reverse an image across the horizontal axis).
+///
+/// See https://uploadcare.com/docs/transformations/image/rotate-flip/#operation-flip
 class FlipTransformation extends NullParamTransformation
     implements ImageTransformation {
   @override
@@ -80,6 +90,8 @@ class FlipTransformation extends NullParamTransformation
 }
 
 /// Mirror an image (mirror-reverse an image across the vertical axis).
+///
+/// See https://uploadcare.com/docs/transformations/image/rotate-flip/#operation-mirror
 class MirrorTransformation extends NullParamTransformation
     implements ImageTransformation {
   @override
@@ -87,6 +99,8 @@ class MirrorTransformation extends NullParamTransformation
 }
 
 /// Desaturate an image.
+///
+/// See https://uploadcare.com/docs/transformations/image/colors/#operation-grayscale
 class GrayscaleTransformation extends NullParamTransformation
     implements ImageTransformation {
   @override
@@ -94,6 +108,8 @@ class GrayscaleTransformation extends NullParamTransformation
 }
 
 /// Invert the colors of an image.
+///
+/// See https://uploadcare.com/docs/transformations/image/colors/#operation-invert
 class InvertTransformation extends NullParamTransformation
     implements ImageTransformation {
   @override
@@ -102,9 +118,11 @@ class InvertTransformation extends NullParamTransformation
 
 /// Auto-enhance an image by performing the following operations: auto levels, auto contrast, and saturation sharpening.
 /// [value] should be in `0..100` range
+///
+/// See https://uploadcare.com/docs/transformations/image/colors/#operation-enhance
 class EnhanceTransformation extends IntTransformation
     implements ImageTransformation {
-  EnhanceTransformation([int value = 50])
+  const EnhanceTransformation([int value = 50])
       : assert(value >= 0 && value <= 100, 'Should be in 0..100 range'),
         super(value);
 
@@ -114,9 +132,11 @@ class EnhanceTransformation extends IntTransformation
 
 /// Sharpen an image, might be especially useful with images that were subjected to downscaling.
 /// [value] should be in `0..20` range
+///
+/// See https://uploadcare.com/docs/transformations/image/blur-sharpen/#operation-sharp
 class SharpTransformation extends IntTransformation
     implements ImageTransformation {
-  SharpTransformation([int value = 5])
+  const SharpTransformation([int value = 5])
       : assert(value >= 0 && value <= 20, 'Should be in 0..20 range'),
         super(value);
 
@@ -128,10 +148,13 @@ class SharpTransformation extends IntTransformation
 /// Technically, strength controls the Gaussian Blur standard deviation multiplied by ten.
 /// Note, different strength values do not affect the operation performance.
 /// [value] should be in `0..5000` range
+///
+/// See https://uploadcare.com/docs/transformations/image/blur-sharpen/#operation-blur
 class BlurTransformation extends IntTransformation
     implements ImageTransformation {
-  BlurTransformation([int value = 5])
-      : assert(value >= 0 && value <= 5000, 'Should be in 0..5000 range'),
+  const BlurTransformation([int? value = 10])
+      : assert(value != null ? value >= 0 && value <= 5000 : true,
+            'Should be in 0..5000 range'),
         super(value);
 
   @override
@@ -140,9 +163,11 @@ class BlurTransformation extends IntTransformation
 
 /// Strip off an ICC profile from an image based on the profile [value] in *kilobytes*.
 /// [value] should be positive [int]
+///
+/// See https://uploadcare.com/docs/transformations/image/colors/#operation-max-icc-size
 class MaxIccSizeTransformation extends IntTransformation
     implements ImageTransformation {
-  MaxIccSizeTransformation([int value = 10])
+  const MaxIccSizeTransformation([int value = 10])
       : assert(value >= 0, 'Should be positive int'),
         super(value);
 
@@ -163,9 +188,11 @@ enum StretchTValue {
 
 /// Set the [ImageResizeTransformation] behavior when provided [Size] are greater than the source image dimensions.
 /// The directive should come before [ImageResizeTransformation] transformation to give any effect.
+///
+/// See https://uploadcare.com/docs/transformations/image/resize-crop/#operation-stretch
 class StretchTransformation extends EnumTransformation<StretchTValue>
     implements ImageTransformation {
-  StretchTransformation([StretchTValue value = StretchTValue.On])
+  const StretchTransformation([StretchTValue value = StretchTValue.On])
       : super(value);
 
   @override
@@ -185,12 +212,14 @@ class StretchTransformation extends EnumTransformation<StretchTValue>
 }
 
 /// Set the fill color when converting alpha channel enabled images to JPEG, with [CropTransformation] or [StretchTransformation].
+///
+/// See https://uploadcare.com/docs/transformations/image/resize-crop/#operation-setfill
 class SetFillTransformation extends Transformation
     implements ImageTransformation {
+  const SetFillTransformation([this.hexColor = '#ffffff']);
+
   /// fill color
   final String hexColor;
-
-  SetFillTransformation([this.hexColor = '#ffffff']);
 
   @override
   String get operation => 'setfill';
@@ -234,8 +263,18 @@ enum ScaleCropTypeTValue {
 
 /// Downscale an image along one of the axes (the one with smaller linear size) and crop the rest.
 /// The method can be implemented with manual, center-focused or “smart crop” behavior.
+///
+/// See https://uploadcare.com/docs/transformations/image/resize-crop/#operation-scale-crop
 class ScaleCropTransformation extends ResizeTransformation
     implements EnumTransformation<ScaleCropTypeTValue>, ImageTransformation {
+  ScaleCropTransformation(
+    Dimensions size, {
+    ScaleCropTypeTValue? type,
+    this.offset,
+    this.center = false,
+  })  : value = type,
+        super(size);
+
   /// an optional [ScaleCropTypeTValue] which enables smart image analysis.
   /// Each smart analysis mode combines various methods for detecting areas of interest in an image.
   /// The methods you include are applied sequentially.
@@ -250,14 +289,6 @@ class ScaleCropTransformation extends ResizeTransformation
   /// centering the crop focus.
   final bool center;
 
-  ScaleCropTransformation(
-    Dimensions size, {
-    ScaleCropTypeTValue? type,
-    this.offset,
-    this.center = false,
-  })  : value = type,
-        super(size);
-
   @override
   String get operation => 'scale_crop';
 
@@ -265,10 +296,7 @@ class ScaleCropTransformation extends ResizeTransformation
   List<String> get params => [
         ...super.params,
         if (value != null) valueAsString,
-        if (offset != null)
-          '${offset!.dx},${offset!.dy}'
-        else if (center)
-          'center',
+        if (offset != null) offset.toString() else if (center) 'center',
       ];
 
   @override
@@ -301,6 +329,8 @@ class ScaleCropTransformation extends ResizeTransformation
 }
 
 /// Resize an image to fit into specified dimensions, proportional.
+///
+/// See https://uploadcare.com/docs/transformations/image/resize-crop/#operation-preview
 class PreviewTransformation extends ResizeTransformation
     implements ImageTransformation {
   PreviewTransformation([Dimensions size = const Dimensions.square(2048)])
@@ -311,19 +341,16 @@ class PreviewTransformation extends ResizeTransformation
 }
 
 /// Crop an image to fit into specified dimensions, implement optional offsets.
+///
+/// See https://uploadcare.com/docs/transformations/image/resize-crop/#operation-crop
 class CropTransformation extends ResizeTransformation
     implements ImageTransformation {
-  /// optional offsets along one or both of the axes
-  final Offsets offset;
-
-  /// centering the crop focus.
-  final bool center;
-
   CropTransformation(
     Dimensions size, [
-    this.offset = Offsets.zero,
-    this.center = false,
+    this.coordinates,
   ]) : super(size);
+
+  final Coordinates? coordinates;
 
   @override
   String get operation => 'crop';
@@ -331,38 +358,23 @@ class CropTransformation extends ResizeTransformation
   @override
   List<String> get params => [
         ...super.params,
-        center ? 'center' : '${offset.dx},${offset.dy}',
+        if (coordinates != null) coordinates.toString(),
       ];
 }
 
 /// Resize an image to fit into specified dimensions.
+///
+/// See https://uploadcare.com/docs/transformations/image/resize-crop/#operation-resize
 class ImageResizeTransformation extends ResizeTransformation
     implements ImageTransformation {
   ImageResizeTransformation(Dimensions size) : super(size);
 }
 
-class OverlayCoordinates {
-  final Offsets offset;
-  final String? predefined;
-
-  const OverlayCoordinates._({
-    this.offset = Offsets.zero,
-    this.predefined,
-  });
-
-  const OverlayCoordinates(this.offset) : predefined = null;
-
-  static const top = OverlayCoordinates._(predefined: 'top');
-  static const bottom = OverlayCoordinates._(predefined: 'bottom');
-  static const left = OverlayCoordinates._(predefined: 'left');
-  static const right = OverlayCoordinates._(predefined: 'right');
-  static const center = OverlayCoordinates._(predefined: 'center');
-
-  @override
-  String toString() => predefined ?? '${offset.dx}p,${offset.dy}p';
-}
-
 /// The overlay operation allows to layer images one over another.
+///
+/// See https://uploadcare.com/docs/transformations/image/overlay/#overlay
+///
+/// You can use overlay transformation to source image, see https://uploadcare.com/docs/transformations/image/overlay/#overlay-self
 ///
 /// Example:
 /// ```dart
@@ -381,20 +393,6 @@ class OverlayCoordinates {
 /// ```
 class OverlayTransformation extends Transformation
     implements ImageTransformation {
-  /// UUID of an image to be layered over input. To be recognized by :uuid, that image should be related to any project of your account.
-  final String imageId;
-
-  /// Linear relative dimensions of the overlay image. The aspect ratio of an overlay is preserved.
-  final Dimensions? dimensions;
-
-  /// Relative position of the overlay over your input. By default, an overlay is positioned in the top-left corner of an input.
-  /// Coordinates represent an offset along each of the axes in either pixel or percent format.
-  /// In general, the coordinate system is similar to the CSS background-position. See [OverlayCoordinates].
-  final OverlayCoordinates? coordinates;
-
-  /// Controls the opacity of the overlay in percent format.
-  final int? opacity;
-
   OverlayTransformation(
     this.imageId, {
     this.dimensions,
@@ -406,13 +404,38 @@ class OverlayTransformation extends Transformation
         assert(coordinates != null ? dimensions != null : true,
             'dimensions should be provided if you are using `coordinates`'),
         assert(
+          dimensions != null ? dimensions.units == MeasureUnits.Percent : true,
+          'Cannot use `MeasureUnits.Pixel` with this transformation',
+        ),
+        assert(
             opacity != null
                 ? dimensions != null &&
                     coordinates != null &&
                     opacity >= 0 &&
                     opacity <= 100
                 : true,
-            '`opacity` should be in 0..100 range');
+            '`opacity` should be in 0..100 range'),
+        assert(
+          coordinates != null && coordinates.predefined == null
+              ? coordinates.offset.units == MeasureUnits.Percent
+              : true,
+          'Cannot use `MeasureUnits.Pixel` in coordinates with this transformation',
+        );
+
+  /// 'self' or UUID of an image to be layered over input. To be recognized by :uuid, that image should be related to any project of your account.
+  /// If you specify 'self' transfomation will be applied to source image
+  final String imageId;
+
+  /// Linear relative dimensions of the overlay image. The aspect ratio of an overlay is preserved.
+  final Dimensions? dimensions;
+
+  /// Relative position of the overlay over your input. By default, an overlay is positioned in the top-left corner of an input.
+  /// Coordinates represent an offset along each of the axes in either pixel or percent format.
+  /// In general, the coordinate system is similar to the CSS background-position. See [Coordinates].
+  final Coordinates? coordinates;
+
+  /// Controls the opacity of the overlay in percent format.
+  final int? opacity;
 
   @override
   String get operation => 'overlay';
@@ -420,7 +443,7 @@ class OverlayTransformation extends Transformation
   @override
   List<String> get params => [
         imageId,
-        if (dimensions != null) '${dimensions!.width}px${dimensions!.height}p',
+        if (dimensions != null) dimensions.toString(),
         if (coordinates != null) coordinates.toString(),
         if (opacity != null) '${opacity}p',
       ];
@@ -440,4 +463,279 @@ class OverlayTransformation extends Transformation
       dimensions.hashCode ^
       coordinates.hashCode ^
       opacity.hashCode;
+}
+
+enum BlurRegionTValue {
+  Region,
+  Faces,
+}
+
+/// Blurs the specified region of the image by the `radius` factor.
+/// The filtering mode is Gaussian Blur, where `radius` parameter sets the blur radius — effect intensity.
+/// Technically, `radius` controls the Gaussian Blur standard deviation multiplied by ten.
+/// The value of `radius` might come up to 5000, while the default value is determined automatically based on the size of the region.
+/// Note, larger `radius` values do not affect the operation performance.
+///
+/// Blur region: https://uploadcare.com/docs/transformations/image/blur-sharpen/#operation-blur-region
+/// Blur faces: https://uploadcare.com/docs/transformations/image/blur-sharpen/#operation-blur-region-faces
+class BlurRegionTransformation extends BlurTransformation {
+  BlurRegionTransformation({
+    this.type = BlurRegionTValue.Region,
+    this.dimensions,
+    this.coordinates,
+    int? radius,
+  })  : assert(
+          type == BlurRegionTValue.Region
+              ? dimensions != null && coordinates != null
+              : true,
+          '`dimensions` and `coordinates` are required for `BlurRegionTValue.Region`',
+        ),
+        super(radius);
+
+  /// When [BlurRegionTValue.Faces] is specified the regions are selected automatically by utilizing face detection.
+  final BlurRegionTValue type;
+  final Dimensions? dimensions;
+  final Offsets? coordinates;
+
+  @override
+  String get operation => 'blur_region';
+
+  @override
+  List<String> get params => [
+        if (type == BlurRegionTValue.Faces) 'faces',
+        if (type == BlurRegionTValue.Region) ...[
+          dimensions!.toString(),
+          coordinates!.toString(),
+        ],
+        ...super.params,
+      ];
+}
+
+/// See: https://uploadcare.com/docs/transformations/image/blur-sharpen/#operation-blur-mask
+class UnsharpMaskingTransformation extends BlurTransformation {
+  const UnsharpMaskingTransformation([this.amount = 100, int radius = 10])
+      : assert(amount >= -200 && amount <= 100, 'Should be in -200..100 range'),
+        super(radius);
+
+  /// - 100 stands for the opaque blur image.
+  /// - 0 stands for no changes in the image, the output is equal to the original.
+  /// - Any negative number would mean subtracting the difference between the blurred and original images from the original. That is the unsharp masking behavior.
+  final int amount;
+
+  @override
+  List<String> get params => [
+        ...super.params,
+        amount.toString(),
+      ];
+}
+
+/// Filter names
+enum FilterTValue {
+  Adaris,
+  Briaril,
+  Calarel,
+  Carris,
+  Cynarel,
+  Cyren,
+  Elmet,
+  Elonni,
+  Enzana,
+  Erydark,
+  Fenralan,
+  Ferand,
+  Galen,
+  Gavin,
+  Gethriel,
+  Iorill,
+  Iothari,
+  Iselva,
+  Jadis,
+  Lavra,
+  Misiara,
+  Namala,
+  Nerion,
+  Nethari,
+  Pamaya,
+  Sarnar,
+  Sedis,
+  Sewen,
+  Sorahel,
+  Sorlen,
+  Tarian,
+  Thellassan,
+  Varriel,
+  Varven,
+  Vevera,
+  Virkas,
+  Yedis,
+  Yllara,
+  Zatvel,
+  Zevcen,
+}
+
+/// Applies one of predefined photo filters by its [FilterTValue].
+/// The way your images look affects their engagement rates.
+/// You apply filters thus providing beautiful images consistent across content pieces you publish.
+///
+/// See https://uploadcare.com/docs/transformations/image/colors/#operation-filter
+class FilterTransformation extends EnumTransformation<FilterTValue>
+    implements ImageTransformation {
+  const FilterTransformation(FilterTValue name, [this.amount = 100])
+      : assert(amount >= -100 && amount <= 200, 'Should be in -100..200 range'),
+        super(name);
+
+  /// can be set in the range from -100 to 200 and defaults to 100. The :amount of:
+  ///
+  /// - 0 stands for no changes in the image, the output is equal to the original.
+  /// - values in the range from 0 to 100 gradually increase filter strength; 100 makes filters work as designed.
+  /// - values over 100 emphasizes filter effect: the strength of applied changes.
+  /// - any negative number would mean subtracting the difference between the filtered and original images from the original. That will produce a "negative filter" effect.
+  final int amount;
+
+  @override
+  String get operation => 'filter';
+
+  @override
+  String get valueAsString {
+    return value.toString().split('.').last.toLowerCase();
+  }
+
+  @override
+  List<String> get params => [
+        ...super.params,
+        amount.toString(),
+      ];
+}
+
+/// Zoom objects operation is best suited for images with solid or uniform backgrounds.
+///
+/// See https://uploadcare.com/docs/transformations/image/resize-crop/#operation-zoom-objects
+class ZoomObjectTransformation extends IntTransformation
+    implements ImageTransformation {
+  const ZoomObjectTransformation(int zoom)
+      : assert(zoom >= 0 && zoom <= 100, 'Should be in 0..100 range'),
+        super(zoom);
+
+  @override
+  String get operation => 'zoom_objects';
+}
+
+/// See https://uploadcare.com/docs/transformations/image/colors/#image-colors-operations
+class ColorBrightnessTransformation extends IntTransformation
+    implements ImageTransformation {
+  const ColorBrightnessTransformation(int value)
+      : assert(value >= -100 && value <= 100, 'Should be in -100..100 range'),
+        super(value);
+
+  @override
+  String get operation => 'brightness';
+}
+
+/// See https://uploadcare.com/docs/transformations/image/colors/#image-colors-operations
+class ColorExposureTransformation extends IntTransformation
+    implements ImageTransformation {
+  const ColorExposureTransformation(int value)
+      : assert(value >= -500 && value <= 500, 'Should be in -500..500 range'),
+        super(value);
+
+  @override
+  String get operation => 'exposure';
+}
+
+/// See https://uploadcare.com/docs/transformations/image/colors/#image-colors-operations
+class ColorGammaTransformation extends IntTransformation
+    implements ImageTransformation {
+  const ColorGammaTransformation(int value)
+      : assert(value >= 0 && value <= 1000, 'Should be in 0..1000 range'),
+        super(value);
+
+  @override
+  String get operation => 'gamma';
+}
+
+/// See https://uploadcare.com/docs/transformations/image/colors/#image-colors-operations
+class ColorContrastTransformation extends IntTransformation
+    implements ImageTransformation {
+  const ColorContrastTransformation(int value)
+      : assert(value >= -100 && value <= 500, 'Should be in -100..500 range'),
+        super(value);
+
+  @override
+  String get operation => 'contrast';
+}
+
+/// See https://uploadcare.com/docs/transformations/image/colors/#image-colors-operations
+class ColorSaturationTransformation extends IntTransformation
+    implements ImageTransformation {
+  const ColorSaturationTransformation(int value)
+      : assert(value >= -100 && value <= 500, 'Should be in -100..500 range'),
+        super(value);
+
+  @override
+  String get operation => 'saturation';
+}
+
+/// See https://uploadcare.com/docs/transformations/image/colors/#image-colors-operations
+class ColorVibranceTransformation extends IntTransformation
+    implements ImageTransformation {
+  const ColorVibranceTransformation(int value)
+      : assert(value >= -100 && value <= 500, 'Should be in -100..500 range'),
+        super(value);
+
+  @override
+  String get operation => 'vibrance';
+}
+
+/// See https://uploadcare.com/docs/transformations/image/colors/#image-colors-operations
+class ColorWarmthTransformation extends IntTransformation
+    implements ImageTransformation {
+  const ColorWarmthTransformation(int value)
+      : assert(value >= -100 && value <= 100, 'Should be in -100..100 range'),
+        super(value);
+
+  @override
+  String get operation => 'warmth';
+}
+
+enum SrgbTValue {
+  Fast,
+  Icc,
+  KeepProfile,
+}
+
+/// The operation sets how Uploadcare behaves depending on different color profiles of uploaded images.
+///
+/// See https://uploadcare.com/docs/transformations/image/colors/#operation-srgb
+class SrgbTransformation extends EnumTransformation<SrgbTValue>
+    implements ImageTransformation {
+  const SrgbTransformation(SrgbTValue value) : super(value);
+
+  @override
+  String get operation => 'srgb';
+
+  @override
+  String get valueAsString {
+    switch (value) {
+      case SrgbTValue.Fast:
+        return 'fast';
+      case SrgbTValue.Icc:
+        return 'icc';
+      case SrgbTValue.KeepProfile:
+        return 'keep_profile';
+      default:
+        return '';
+    }
+  }
+}
+
+/// By default, CDN instructs browsers to show images and download other file types.
+/// The inline control allows you to change this behavior.
+///
+/// See https://uploadcare.com/docs/delivery/cdn/#inline
+class InlineTransformation extends BooleanTransformation
+    implements ImageTransformation {
+  const InlineTransformation(bool inline) : super(inline);
+
+  @override
+  String get operation => 'inline';
 }