+++ uploadcare_flutter/CHANGELOG.md
@@ -1,12 +1,12 @@
-## [1.0.2] - Sat Mar 6 2021
+## [2.0.0] - Sat Mar 6 2021
+* Updated `uploadcare_client` dependency with Breaking changes
 
+## [1.0.2] - Sat Mar 6 2021
 * Updated `uploadcare_client` dependency 
 
 ## [1.0.1] - Sat Mar 6 2021
-
 * Updated `uploadcare_client` dependency 
 
 ## [1.0.0] - Sat Mar 6 2021
-
 * Initial release.
 * Added measures and face entities extensions for base package `uploadcare_client@^3.0.0`